converted to double precision - third build - succesful simulation
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@1865 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ MODFLOW/MF2005/gwf2met1.f
@@ -513,8 +513,8 @@ c arguments
       integer, intent(in) :: igrid
       integer, intent(in) :: ncol
       integer, intent(in) :: nrow
-      real, dimension(ncol), intent(in) :: delr
-      real, dimension(nrow), intent(in) :: delc
+      real(kind=8), dimension(ncol), intent(in) :: delr
+      real(kind=8), dimension(nrow), intent(in) :: delc
 
 c local variables
       integer :: icol, irow
